+++ lib/inc/lqr_pid_control.h
@@ -2,23 +2,23 @@
 
 typedef struct pid_parameter
 {
-    double p_gain;
-    double i_gain;
-    double d_gain;
-    double target;
-    double error;
-    double pri_error;
-    double error_limit;
-    double error_sum;
-    double error_sum_limit;
-    double output;
+    float p_gain;
+    float i_gain;
+    float d_gain;
+    float target;
+    float error;
+    float pri_error;
+    float error_limit;
+    float error_sum;
+    float error_sum_limit;
+    float output;
 } pid_parameter_t;
 
 class pid_controller
 {
     public:
         pid_controller();
-        pid_controller(double p_gain, double i_gain, double d_gain)
+        pid_controller(float p_gain, float i_gain, float d_gain)
         {
             init(p_gain, i_gain, d_gain);
         }
@@ -26,19 +26,19 @@ class pid_controller
         {
             this->parameter = parameter;
         }
-        void init(double p_gain, double i_gain, double d_gain);
-        void set_gain(double p_gain, double i_gain, double d_gain);
-        void set_target(double target);
-        double get_p_gain() {
+        void init(float p_gain, float i_gain, float d_gain);
+        void set_gain(float p_gain, float i_gain, float d_gain);
+        void set_target(float target);
+        float get_p_gain() {
             return this->parameter.p_gain;
         };
-        double get_i_gain() {
+        float get_i_gain() {
             return this->parameter.i_gain;
         };
-        double get_d_gain() {
+        float get_d_gain() {
             return this->parameter.d_gain;
         };
-        double calculate(double current_value);
+        float calculate(float current_value);
 
     private:
         pid_parameter_t parameter;
